% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_rate.R
\name{boxplot}
\alias{boxplot}
\title{Get five summary numbers for ramen in different countries}
\usage{
boxplot(continent = NULL)
}
\arguments{
\item{continent}{The continent in which a country is located in.}
}
\value{
A boxplot contains 5 summary statistics of different ramen.
}
\description{
This function gets the average ratings for ramen based on countries.
}
\examples{

boxplot("Asia")


}
