% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_numeric.R
\name{five_sum}
\alias{five_sum}
\title{Get five summary numbers for different ramen styles}
\usage{
five_sum(brand = NULL)
}
\arguments{
\item{brand}{The ramen brand}
}
\value{
A data table contains 5 summary statistics.
}
\description{
This function gets the five number summary for selected ramen styles.
}
