% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_rate.R
\name{get_average}
\alias{get_average}
\title{Get the average ratings}
\usage{
get_average(continents)
}
\arguments{
\item{continents}{The continent in which the country is located.}
}
\value{
A interactive boxplot shows the five summary values for each country.
}
\description{
This function produces graphs showing averate ratings on each country, across continents
}
\examples{
get_pct(names = "Australia")
get_pct(names = "Japan")
}
