% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_numeric.R
\name{get_best_style}
\alias{get_best_style}
\title{Get highest varieties}
\usage{
get_best_style(styles)
}
\arguments{
\item{styles}{The style of ramen containers}
}
\value{
A sentence describes the number of variations.
}
\description{
This function gets the highest varietie.
}
