% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_rate.R
\name{get_pct}
\alias{get_pct}
\title{Get good and bad ramen from each country}
\usage{
get_pct(names)
}
\arguments{
\item{names}{The country name in which ramen comes from.}
}
\value{
A pie chart shows the corresponding percentage of ramen rated as either good or bad.
}
\description{
This function generates plots illustrating the proportion of good and bad ramen for each nation.
}
