% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ramen_review.R
\name{get_ramen_rating}
\alias{get_ramen_rating}
\title{@title
Get ramen_rating dataset}
\usage{
get_ramen_rating()
}
\value{
A data frame including 3180 rows and 7 columns.
}
\description{
This function gets the ramen_rating dataset
}
