% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_numeric.R
\name{get_rank}
\alias{get_rank}
\title{Rank highest average rate}
\usage{
get_rank(styles)
}
\arguments{
\item{styles}{The style of ramen containers}
}
\value{
A sentence ranks the average rate .
}
\description{
This function gets the ranking of average rate for each ramen style.
}
