% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ramen_review.R
\name{get_table}
\alias{get_table}
\title{Get ramen_rating dataset in tabular format}
\usage{
get_table()
}
\value{
A tidy table comprises of 38 rows and 5 columns.
}
\description{
This function provides 4 summary statistics for ramen rating dataset
}
