% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_rate.R
\name{good_bad_ramen}
\alias{good_bad_ramen}
\title{Get five summary numbers for ramen in different countries}
\usage{
good_bad_ramen(country)
}
\arguments{
\item{country}{The country in which a ramen bran comes from.}
}
\value{
A boxplot contains 5 summary statistics of different ramen.
}
\description{
This function gets the average ratings for ramen based on countries.
}
\examples{
good_bad_ramen("Australia")


}
