\name{ramen_rating}
\alias{ramen_rating}
\docType{data}
\title{Raw ramen ratings dataset}
\description{Dataset provides reviews on more than 200 ramen varieties in selected countries}
\usage{data("ramen_rating")}

\format{A data frame with 3180 observations and 6 variables.}

\references{https://www.theramenrater.com/resources-2/the-list/}

\examples{
data(ramen_rating.R)}
\keyword{datasets}
