% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ramen_review.R
\name{ramen_ratings}
\alias{ramen_ratings}
\title{Get review ratings for different ramen varieties}
\usage{
ramen_ratings()
}
\value{
A data.frame including 2805 observations and 7 variables.
}
\description{
This function gets the ratings for various ramen
}
\examples{
ramen_ratings()

}
